requirements.txt
beautifulsoup4==4.12.2
google==3.0.0
grpcio==1.54.2
install==1.3.5
protobuf==4.23.2
pyfiglet==0.8.post1
soupsieve==2.4.1
